pyproject.toml
[project]
name = "dataspine"
version = "0.1.0"
description = "Institutional data operations spine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "apscheduler>=3.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "smoke: smoke tests",
    "integration: integration tests",
    "slow: slow running tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "UP017",  # datetime.UTC alias (requires Python 3.11+, keep timezone.utc for compatibility)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
